Prettify Auto PR workflow
The automated PRs are not too pretty to look at atm and also includes a unclickable link to PRs. 

This PR changes the url used to html_url which is the one used when reading PRs from the browser.
Further more it uses markdown formatting instead of a code block.